+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 6/16/2023
 - Minor: Discord embed generator now accept `stamina` and `expedition` as a parameter to generate the embed
+- Minor: Added missing argument handler for `--sign` and `--stamina` flag
 
 ## 6/15/2023
 - Major: Moved `Discord`, `Telegram` to controllers folder